[pdf2vqa] 一个文件一个算子
